chore: refactor based on review comments
