+++ packages/benchmark/package.json
@@ -23,10 +23,10 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@agoric/assert": "^0.6.0",
+    "@agoric/boot": "^0.1.0",
     "@agoric/cosmic-swingset": "^0.41.3",
     "@agoric/internal": "^0.3.2",
     "@agoric/inter-protocol": "^0.16.1",
-    "@agoric/test-support": "^0.1.0",
     "@agoric/vats": "^0.15.1",
     "@agoric/zoe": "^0.26.2",
     "@endo/init": "^0.5.59"
