+++ packages/boot/package.json
@@ -9,7 +9,6 @@
     "build": "exit 0",
     "test": "ava",
     "test:xs": "SWINGSET_WORKER_TYPE=xs-worker ava 'test/bootstrapTests/**/test-*.js' 'test/upgrading/**/test-*.js'",
-    "bench": "ava --config ./ava.bench.config.js",
     "lint-fix": "yarn lint:eslint --fix",
     "lint": "run-s --continue-on-error lint:*",
     "lint:types": "tsc",
@@ -19,9 +18,16 @@
   "author": "Agoric",
   "license": "Apache-2.0",
   "dependencies": {
+    "@agoric/assert": "^0.6.0",
     "@agoric/builders": "^0.1.0",
+    "@agoric/cosmic-swingset": "^0.41.3",
     "@agoric/ertp": "^0.16.2",
     "@agoric/internal": "^0.3.2",
+    "@agoric/inter-protocol": "^0.16.1",
+    "@agoric/kmarshal": "^0.1.0",
+    "@agoric/swing-store": "^0.9.1",
+    "@agoric/swingset-vat": "^0.32.2",
+    "@agoric/time": "^0.3.2",
     "@agoric/vat-data": "^0.5.2",
     "@agoric/vats": "^0.15.1",
     "@agoric/vm-config": "^0.1.0",
@@ -37,19 +43,10 @@
     "import-meta-resolve": "^2.2.1"
   },
   "devDependencies": {
-    "@agoric/assert": "^0.6.0",
-    "@agoric/benchmark": "^0.1.0",
-    "@agoric/cosmic-swingset": "^0.41.3",
     "@agoric/deploy-script-support": "^0.10.3",
     "@agoric/governance": "^0.10.3",
-    "@agoric/inter-protocol": "^0.16.1",
-    "@agoric/kmarshal": "^0.1.0",
     "@agoric/store": "^0.9.2",
-    "@agoric/swing-store": "^0.9.1",
     "@agoric/swingset-liveslots": "^0.10.2",
-    "@agoric/swingset-vat": "^0.32.2",
-    "@agoric/test-support": "^0.1.0",
-    "@agoric/time": "^0.3.2",
     "ava": "^5.3.0",
     "c8": "^7.13.0",
     "tsx": "^3.12.8"
