+++ packages/boot/test/bootstrapTests/liquidation.ts
@@ -8,13 +8,13 @@ import {
   makeAgoricNamesRemotesFromFakeStorage,
 } from '@agoric/vats/tools/board-utils.js';
 import { Offers } from '@agoric/inter-protocol/src/clientSupport.js';
-import { ExecutionContext } from 'ava';
+import type { ExecutionContext } from 'ava';
+import { makeSwingsetTestKit } from '../../tools/supports.ts';
 import {
   makeGovernanceDriver,
   makePriceFeedDriver,
   makeWalletFactoryDriver,
-  makeSwingsetTestKit,
-} from '@agoric/test-support';
+} from '../../tools/drivers.ts';
 
 export type LiquidationSetup = {
   vaults: {
