+++ packages/boot/test/bootstrapTests/test-vaults-integration.ts
@@ -13,10 +13,8 @@ import {
 } from '@agoric/vats/tools/board-utils.js';
 import type { TestFn } from 'ava';
 import { ParamChangesOfferArgs } from '@agoric/inter-protocol/src/econCommitteeCharter.js';
-import {
-  makeSwingsetTestKit,
-  makeWalletFactoryDriver,
-} from '@agoric/test-support';
+import { makeSwingsetTestKit } from '../../tools/supports.ts';
+import { makeWalletFactoryDriver } from '../../tools/drivers.ts';
 
 // presently all these tests use one collateral manager
 const collateralBrandKey = 'ATOM';
