+++ packages/boot/test/bootstrapTests/test-vaults-upgrade.ts
@@ -15,10 +15,8 @@ import { makeAgoricNamesRemotesFromFakeStorage } from '@agoric/vats/tools/board-
 import { ExecutionContext, TestFn } from 'ava';
 import { FakeStorageKit } from '@agoric/internal/src/storage-test-utils.js';
 import { EconomyBootstrapSpace } from '@agoric/inter-protocol/src/proposals/econ-behaviors.js';
-import {
-  makeSwingsetTestKit,
-  makeWalletFactoryDriver,
-} from '@agoric/test-support';
+import { makeSwingsetTestKit } from '../../tools/supports.ts';
+import { makeWalletFactoryDriver } from '../../tools/drivers.ts';
 
 // presently all these tests use one collateral manager
 const collateralBrandKey = 'ATOM';
