+++ packages/boot/tools/drivers.ts
@@ -22,7 +22,8 @@ import type { WalletFactoryStartResult } from '@agoric/vats/src/core/startWallet
 import type { OfferSpec } from '@agoric/smart-wallet/src/offers.js';
 import type { TimerService } from '@agoric/time/src/types.js';
 import type { OfferMaker } from '@agoric/smart-wallet/src/types.js';
-import type { RunUtils, SwingsetTestKit } from './supports.ts';
+import type { RunUtils } from '@agoric/swingset-vat/tools/run-utils.ts';
+import type { SwingsetTestKit } from './supports.ts';
 
 export const makeWalletFactoryDriver = async (
   runUtils: RunUtils,
