--- packages/test-support/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@agoric/test-support",
-  "version": "0.1.0",
-  "description": "Support libraries for testing on chain",
-  "type": "module",
-  "main": "./src/index.js",
-  "exports": {
-    ".": "./src/index.js"
-  },
-  "repository": "https://github.com/Agoric/agoric-sdk",
-  "author": "Agoric",
-  "license": "Apache-2.0",
-  "scripts": {
-    "build": "exit 0",
-    "test": "exit 0",
-    "test:xs": "exit 0",
-    "lint-fix": "yarn lint:eslint --fix",
-    "lint": "run-s --continue-on-error lint:*",
-    "lint:types": "tsc",
-    "lint:eslint": "eslint ."
-  },
-  "dependencies": {
-    "@endo/eventual-send": "^0.17.5",
-    "@endo/stream": "^0.3.28",
-    "@agoric/assert": "^0.6.0",
-    "@agoric/cosmic-swingset": "^0.41.3",
-    "@agoric/inter-protocol": "^0.16.1",
-    "@agoric/internal": "^0.3.2",
-    "@agoric/kmarshal": "^0.1.0",
-    "@agoric/smart-wallet": "^0.5.3",
-    "@agoric/swing-store": "^0.9.1",
-    "@agoric/swingset-vat": "^0.32.2",
-    "@agoric/time": "^0.3.2",
-    "@agoric/vats": "^0.15.1",
-    "ava": "^5.3.0",
-    "import-meta-resolve": "^2.2.1"
-  },
-  "devDependencies": {},
-  "publishConfig": {
-    "access": "public"
-  },
-  "ava": {
-    "files": [
-      "test/**/test-*.js"
-    ],
-    "require": [
-      "@endo/init/debug.js"
-    ],
-    "timeout": "2m"
-  }
-}
