--- packages/test-support/src/index.js
@@ -1,2 +1,0 @@
-export * from './supports.ts';
-export * from './drivers.ts';
