--- packages/test-support/tsconfig.json
@@ -1,14 +1,0 @@
-// This file can contain .js-specific Typescript compiler config.
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "allowImportingTsExtensions": true,
-    "allowSyntheticDefaultImports": true,
-    "checkJs": true,
-    "maxNodeModuleJsDepth": 2,
-  },
-  "include": [
-    "src/**/*.js",
-    "test/**/*.js"
-  ]
-}
